Add raster data from geodataset (#20)
* add raster data from geodataset

* add technical docs

* changelog and minor review comment
