+++ docs/api/advanced/workflows/grid.rst
@@ -1,0 +1,14 @@
+.. currentmodule:: hydromt_wflow.workflows.grid
+
+.. _api_workflows_grid:
+
+
+====
+Grid
+====
+
+.. autosummary::
+    :toctree: _generated
+    :nosignatures:
+
+    grid_from_geodataset
