+++ docs/changelog.rst
@@ -6,12 +6,24 @@ All notable changes to this project will be documented in this page.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
-v1.0.2 (3 March 2026)
-==========================
+Unreleased
+==========
 Added
 -----
 - **setup_agroforestry**: method to add agroforestry areas to landuse maps and update related vegetation parameters. PR #2
 - **setup_subbasins**: method to delineate subbasins based on streamorder, pfafstetter or area threshold. PR #7
+- **setup_grid_from_geodataset**: method to set static/cyclic/forcing grid data from a geodataset. PR #20
+
+Changed
+-------
+
+Fixed
+-----
+
+v1.0.2 (3 March 2026)
+=====================
+Added
+-----
 - Added Pandas 3.0 support (#713)
 
 Fixed
@@ -19,7 +31,7 @@ Fixed
 - ``workflows.demand.surfacewaterfrac_used`` synchronize nodata value and fill value to prevent skipping interpolation. (#706)
 
 v1.0.1 (8 January 2026)
-==========================
+=======================
 Added
 -----
 - Artifact test automation in release pipelines (#690)
@@ -30,7 +42,7 @@ Fixed
 
 
 v1.0.0 (18 December 2025)
-==========================
+=========================
 Overview
 --------
 Version 1.0.0 is a major release that introduces full support for Wflow.jl 1.0.0 and drops support for all earlier Wflow.jl versions.
@@ -39,27 +51,27 @@ An explicit upgrade path is provided via the upgrade_to_v1_wflow function to hel
 For more information, see the [migration guide](https://deltares.github.io/hydromt_wflow/latest/user_guide/6_migration_guide/index.html)
 
 Architecture and Model Structure
------------------------------------
+--------------------------------
 The internal model structure has been refactored around hydromt-core v1 concepts.
 The model root is now managed by a ModelRoot class, and configuration, static maps, forcing, states, geometries, and outputs are implemented as ModelComponent classes with a uniform data access pattern.
 Component and method names were harmonized with Wflow.jl conventions, including renaming grid to staticmaps and standardizing static map, state, and geometry names.
 
 Configuration and Setup Changes
------------------------------------
+-------------------------------
 TOML configuration handling was redesigned to match Wflow.jl 1.0.0.
 Configuration options are now added incrementally by setup methods rather than generated from a full template.
 Several setup methods were renamed or split for clarity, and clipping a model is now handled through a single clip method.
 Users can explicitly control variable names in static maps and selected geometries, enabling multiple variants of the same parameter within one model.
 
 Lakes, Reservoirs, and Rivers
------------------------------------
+-----------------------------
 Lakes and reservoirs are now handled through a unified reservoir framework.
 Reservoirs can be added or overwritten, and a combined reservoir geometry is generated by default.
 River setup logic was split into generic and model-specific parts, improving reuse and maintainability.
 Sediment-related workflows were updated to prepare only sediment-relevant variables.
 
 Compatibility, Fixes, and Cleanup
------------------------------------
+---------------------------------
 The minimum supported Python version was raised to 3.11.
 Legacy functionality tied to older Wflow versions and deprecated dependencies was removed, including pcraster support and model-level get/set/read/write methods.
 Numerous bug fixes and behavioral improvements were included, particularly around upgrading models, reservoir handling, forcing generation, and geometry processing.
