+++ docs/user_guide/2_sbm_model/1_methods_components.rst
@@ -269,6 +269,9 @@ Additional high-level utilities to modify model geometry, link external models, 
     * - :py:meth:`~WflowSbmModel.setup_grid_from_raster`
       -  Setup staticmaps from raster to add parameters from direct data.
       - :py:meth:`~WflowSbmModel.setup_basemaps`
+    * - :py:meth:`~WflowSbmModel.setup_grid_from_geodataset`
+      -  Setup static/cyclic/forcing from geodataset to add parameters from direct data.
+      - :py:meth:`~WflowSbmModel.setup_basemaps`
     * - :py:meth:`~WflowSbmModel.upgrade_to_v1_wflow`
       -  Upgrade a model from a Wflow.jl 0.x to 1.0 .
       -
