+++ docs/user_guide/3_sediment_model/1_methods_components.rst
@@ -157,6 +157,9 @@ Other methods
     * - :py:meth:`~WflowSedimentModel.setup_grid_from_raster`
       -  Setup staticmaps from raster to add parameters from direct data.
       - :py:meth:`~WflowSedimentModel.setup_basemaps`
+    * - :py:meth:`~WflowSbmModel.setup_grid_from_geodataset`
+      -  Setup static/cyclic/forcing from geodataset to add parameters from direct data.
+      - :py:meth:`~WflowSbmModel.setup_basemaps`
     * - :py:meth:`~WflowSedimentModel.upgrade_to_v1_wflow`
       -  Upgrade a model from a Wflow.jl 0.x to 1.0 .
       -
