+++ docs/user_guide/5_setup_methods/index.rst
@@ -29,3 +29,4 @@ separately in the **User Guide**.
    setup_x
    setup_lulcmaps
    setup_agroforestry
+   setup_grid_from_geodataset
