+++ hydromt_wflow/naming.py
@@ -459,6 +459,16 @@ WFLOW_NAMES: dict[str, dict[str, str | None]] = {
         "wflow_v1": "reservoir_water__target_min_volume_fraction",
         "hydromt_name": "reservoir_target_min_fraction",
     },
+    "reservoir_outflow": {
+        "wflow_v0": None,
+        "wflow_v1": "reservoir_water__outgoing_observed_volume_flow_rate",
+        "hydromt_name": "reservoir_outflow",
+    },
+    "reservoir_external_inflow": {
+        "wflow_v0": None,
+        "wflow_v1": "reservoir_water__external_inflow_volume_flow_rate",
+        "hydromt_name": "reservoir_external_inflow",
+    },
     # gwf
     "altitude": {
         "wflow_v0": "altitude",
