+++ hydromt_wflow/wflow_base.py
@@ -16,6 +16,7 @@ import xarray as xr
 from hydromt import hydromt_step
 from hydromt.error import NoDataStrategy
 from hydromt.gis import flw
+from hydromt.gis.vector import GeoDataArray
 from hydromt.model import Model
 
 import hydromt_wflow.utils as utils
@@ -1618,6 +1619,191 @@ one variable and variables list is not provided."
                     self.config.set(f"input.static.{wflow_variables[i]}", variables[i])
 
     @hydromt_step
+    def setup_grid_from_geodataset(
+        self,
+        geodataset_fn: str | pd.DataFrame | xr.Dataset,
+        locations_fn: str | gpd.GeoDataFrame | None = None,
+        index_col: str | None = None,
+        variable: str | None = None,
+        fill_value: float | int | None = None,
+        nodata_value: float | int = -9999,
+        mask: str | None = None,
+        output_names: dict = {},
+        resample_time_kwargs: dict | None = None,
+    ) -> None:
+        """
+        Add data variable from ``geodataset_fn`` to grid object.
+
+        Prepares a static/cyclic/forcing grid layer at specificied point locations.
+        A GeoDataset including the locations and the variable to rasterize can be used.
+        Alternatively, locations can be provided as a separate geodataframe or as a
+        layer already present in the staticmaps (e.g. gauges locations or reservoir
+        outlets).
+
+        Locations will be used as is. If snapping is required, users can first snap
+        using the ``setup_gauges`` method.
+
+        The config toml can also be updated to include the new map using
+        ``wflow_variable``.
+
+        Adds model layers:
+
+        * **variable** grid: data from geodataset_fn
+
+        Parameters
+        ----------
+        geodataset_fn : str, pd.DataFrame, xr.Dataset
+            Geodataset source as DataFrame or GeoDataset.
+
+            - DataFrame: the index column should contain time and the other
+              columns should correspond to the name or ID values of the stations
+              in `locations_fn`.
+            - GeoDataset: the dataset should contain the variable specified by
+              `variable` and the dimensions 'time' and 'index'.
+
+            * Required variable: 'time', 'variable'
+
+        locations_fn : str, gpd.GeoDataFrame, optional
+            Source for the locations of the stations as points: (x, y) or (lat, lon).
+            Only required if geodataset_fn is of type DataFrame. Can be provided as
+            a layer already present in the staticmaps.
+        index_col : str, optional
+            Column in locations_fn to use for station ID values, by default None.
+        variable : str, optional
+            Name of the variable in geodataset_fn to add to the grid. Only required if
+            geodataset_fn is a GeoDataset with multiple variables.
+        fill_value : float | int, optional
+            Fill value for the other grid cells. If not provided, the same value as the
+            nodata value is used. By default None.
+        nodata_value : float | int, optional
+            No data value for the grid cells if not defined in the geodataset.
+        mask : str, optional
+            Name of the mask to apply on the grid. Should be a layer already present in
+            the staticmaps. If None, the basin mask will be used.
+        output_names : dict, optional
+            Dictionary with output name that will be used in the model netcdf input
+            files. Users should provide the Wflow.jl variable name followed by the name
+            in the netcdf file.
+        resample_time_kwargs : dict, optional
+            If time resampling is required in case of forcing data, the kwargs to pass
+            to the hydromt.model.processes.meteo.resample_time method can be specified
+            here. For example ``upsampling`` (bfill), ``downsampling`` (mean) or
+            ``conserve_mass`` (True).
+        """
+        logger.info(f"Preparing grid data from geodataset source {geodataset_fn}")
+        self._update_naming(output_names)
+
+        # In case data type is forcing define time_range to read in data here
+        starttime = self.config.get_value("time.starttime")
+        endtime = self.config.get_value("time.endtime")
+        freq = pd.to_timedelta(self.config.get_value("time.timestepsecs"), unit="s")
+
+        # Check data type of geodataset_fn if it is provided through the data catalog
+        if isinstance(geodataset_fn, str) and geodataset_fn in self.data_catalog:
+            _data_type = self.data_catalog.get_source(geodataset_fn).data_type
+        else:
+            _data_type = None
+
+        # Read geodataset data
+        if (
+            isinstance(geodataset_fn, xr.Dataset | xr.DataArray)
+            or _data_type == "GeoDataset"
+        ):
+            da = self.data_catalog.get_geodataset(
+                geodataset_fn,
+                geom=self.region,
+                time_range=(starttime, endtime),
+                variables=[variable] if variable is not None else None,
+                single_var_as_array=True,
+            )
+        else:
+            # Read timeseries
+            df = self.data_catalog.get_dataframe(
+                geodataset_fn,
+                time_range=(starttime, endtime),
+            )
+            # Get locs
+            if locations_fn is None:
+                raise ValueError(
+                    "Using a DataFrame as geodataset source requires that "
+                    "locations are provided separately through locations_fn."
+                )
+            elif (
+                isinstance(locations_fn, str)
+                and locations_fn in self.staticmaps.data.data_vars
+            ):
+                locs = self.staticmaps.data[locations_fn]
+                gdf_locs = locs.raster.vectorize()
+                centroid = utils.planar_operation_in_utm(
+                    gdf_locs, lambda geom: geom.centroid
+                )
+                gdf_locs["geometry"] = centroid
+                # After rasterize use col value for index and reapply dtype
+                gdf_locs.index = gdf_locs["value"].astype(locs.dtype)
+
+            else:
+                # Load the locations
+                gdf_locs = self.data_catalog.get_geodataframe(
+                    locations_fn,
+                    geom=self.region,
+                    # assert_gtype= "Point", hydromt#1243
+                    handle_nodata=NoDataStrategy.IGNORE,
+                )
+                # In case no locations within basin
+                if gdf_locs.empty:
+                    logger.warning(
+                        f"No shapes of {locations_fn} found within region, "
+                        "skipping method."
+                    )
+                    return
+                # Use station ids from gdf_stations when reading the DataFrame
+                if index_col is not None:
+                    gdf_locs = gdf_locs.set_index(index_col)
+
+            # Index is required to contruct GeoDataArray
+            if gdf_locs.index.name is None:
+                gdf_locs.index.name = "stations"
+
+            # Convert to geodataset
+            da = GeoDataArray.from_gdf(
+                gdf=gdf_locs,
+                data=df,
+                name=variable,
+                index_dim=None,
+                dims=["time", gdf_locs.index.name],
+                keep_cols=False,
+                merge_index="inner",
+            )
+
+        # Prepare the geodataset data
+        da_out, da_type = workflows.grid.grid_from_geodataset(
+            da=da,
+            ds_like=self.staticmaps.data,
+            fill_value=fill_value,
+            nodata_value=nodata_value,
+            mask=mask,
+            freq=freq,
+            resample_time_kwargs=resample_time_kwargs,
+        )
+
+        # Update toml
+        if any(output_names):
+            if len(output_names) > 1:
+                raise ValueError(
+                    "Only one output name is allowed for areamap, \
+                    please provide a dictionary with one key."
+                )
+            variable_name = list(output_names.values())[0]
+            da_out = da_out.rename(variable_name)
+            self._update_config_variable_name(variable_name, data_type=da_type)
+
+        # Add to grid or forcing
+        if da_type == "forcing":
+            self.forcing.set(da_out)
+        else:  # cyclic or static
+            self.staticmaps.set(da_out)
+
+    @hydromt_step
     def setup_areamap(
         self,
         area_fn: str | gpd.GeoDataFrame,
@@ -1952,14 +2138,23 @@ one variable and variables list is not provided."
         data_type: str, optional
             Type of data (static, forcing, cyclic, None), by default "static"
         """
+        data_types = ["static", "forcing", "cyclic", None]
+
         data_vars = [data_vars] if isinstance(data_vars, str) else data_vars
         _prefix = f"input.{data_type}" if data_type is not None else "input"
+        _other_prefixes = [
+            f"input.{dt}" for dt in data_types if dt != data_type and dt is not None
+        ]
+
         for var in data_vars:
             if var in self._WFLOW_NAMES:
                 # Get the name from the Wflow variable name
                 wflow_var = self._WFLOW_NAMES[var]
                 # Update the config variable name
                 self.config.set(f"{_prefix}.{wflow_var}", var)
+                # Remove from other prefixes if present
+                for other_prefix in _other_prefixes:
+                    self.config.remove(f"{other_prefix}.{wflow_var}", errors="ignore")
             # else not a wflow variable
             # (spelling mistakes should have been checked in _update_naming)
