+++ hydromt_wflow/workflows/__init__.py
@@ -6,6 +6,7 @@ from hydromt_wflow.workflows.demand import *
 from hydromt_wflow.workflows.forcing import *
 from hydromt_wflow.workflows.gauges import *
 from hydromt_wflow.workflows.glaciers import *
+from hydromt_wflow.workflows.grid import *
 from hydromt_wflow.workflows.landuse import *
 from hydromt_wflow.workflows.reservoirs import *
 from hydromt_wflow.workflows.river import *
